migrate to 4.9 - package index, codec/lucene40, codec/lucene42

+++ core/index/postings.go
@@ -1313,7 +1313,7 @@ func (f *segmentTermsEnumFrame) scanToTermLeaf(target []byte, exactOnly bool) (s
 
 		termLen := f.prefix + f.suffix
 		f.startBytePos = f.suffixesReader.Pos
-		f.suffixesReader.SkipBytes(f.suffix)
+		f.suffixesReader.SkipBytes(int64(f.suffix))
 
 		targetLimit := termLen
 		if len(target) < termLen {
